refactor: Extract bot core logic to its own module
