+++ src/bot-core/reactions.js
